+++ library/convert_type_k.h
@@ -1,3 +1,3 @@
 
-float typeK_temp_to_mv(float t90);
-float typeK_mv_to_temp(float mv);
+double typeK_temp_to_mv(double t90);
+double typeK_mv_to_temp(double mv);
